+++ internal/pkg/api/vsphere.go
@@ -164,13 +164,6 @@ func WithDatacenter(value string) VSphereFiller {
 	}
 }
 
-// WithDisableCSI sets the value for DisableCSI in VSphereDatacenterConfig.
-func WithDisableCSI(value bool) VSphereFiller {
-	return func(config VSphereConfig) {
-		config.datacenterConfig.Spec.DisableCSI = value
-	}
-}
-
 // WithCloneModeForAllMachines sets the CloneMode for all VSphereMachineConfigs.
 func WithCloneModeForAllMachines(value anywherev1.CloneMode) VSphereFiller {
 	return func(config VSphereConfig) {