+++ pkg/config/vsphereuser.go
@@ -12,9 +12,6 @@ const (
 	EksavSphereCPUsernameKey = "EKSA_VSPHERE_CP_USERNAME"
 	// EksavSphereCPPasswordKey holds Password for cloud provider.
 	EksavSphereCPPasswordKey = "EKSA_VSPHERE_CP_PASSWORD"
-	// EksavSphereCSIUsernameKey holds Username and password for the CSI driver.
-	EksavSphereCSIUsernameKey = "EKSA_VSPHERE_CSI_USERNAME"
-	EksavSphereCSIPasswordKey = "EKSA_VSPHERE_CSI_PASSWORD"
 )
 
 type VSphereUserConfig struct {
@@ -22,8 +19,6 @@ type VSphereUserConfig struct {
 	EksaVspherePassword    string
 	EksaVsphereCPUsername  string
 	EksaVsphereCPPassword  string
-	EksaVsphereCSIUsername string
-	EksaVsphereCSIPassword string
 }
 
 //go:embed static/globalPrivs.json
@@ -34,18 +29,6 @@ var VSphereUserPrivsFile string
 
 //go:embed static/adminPrivs.json
 var VSphereAdminPrivsFile string
-
-//go:embed static/cnsDatastorePrivs.json
-var VSphereCnsDatastorePrivsFile string
-
-//go:embed static/cnsSearchAndSpbmPrivs.json
-var VSphereCnsSearchAndSpbmPrivsFile string
-
-//go:embed static/cnsVmPrivs.json
-var VSphereCnsVmPrivsFile string
-
-//go:embed static/cnsHostConfigStorage.json
-var VSphereCnsHostConfigStorageFile string
 
 //go:embed static/readOnlyPrivs.json
 var VSphereReadOnlyPrivs string
@@ -62,21 +45,12 @@ func NewVsphereUserConfig() *VSphereUserConfig {
 		eksaCPUsername = eksaVsphereUsername
 		eksaCPPassword = eksaVspherePassword
 	}
-	// CSI driver credentials
-	eksaCSIUsername := os.Getenv(EksavSphereCSIUsernameKey)
-	eksaCSIPassword := os.Getenv(EksavSphereCSIPasswordKey)
-	if eksaCSIUsername == "" {
-		eksaCSIUsername = eksaVsphereUsername
-		eksaCSIPassword = eksaVspherePassword
-	}
 
 	vuc := VSphereUserConfig{
 		eksaVsphereUsername,
 		eksaVspherePassword,
 		eksaCPUsername,
 		eksaCPPassword,
-		eksaCSIUsername,
-		eksaCSIPassword,
 	}
 
 	return &vuc