+++ release/api/v1alpha1/bundle_types.go
@@ -193,10 +193,6 @@ type VSphereBundle struct {
 	KubeProxy            Image    `json:"kubeProxy"`
 	Manager              Image    `json:"manager"`
 	KubeVip              Image    `json:"kubeVip"`
-	// This field has been deprecated
-	Driver               Image    `json:"driver,omitempty"`
-	// This field has been deprecated
-	Syncer               Image    `json:"syncer,omitempty"`
 	Components           Manifest `json:"components"`
 	Metadata             Manifest `json:"metadata"`
 	ClusterTemplate      Manifest `json:"clusterTemplate"`